+++ easy-sys/src/main/java/com/easy/admin/util/SysConfigUtil.java
@@ -56,8 +56,8 @@ public class SysConfigUtil {
         if (config != null) {
             if (DataTypeConst.STRING.equals(config.getType())) {
                 return config.getValue();
-            } else if (DataTypeConst.INTEGER.equals(config.getType())) {
-                return Integer.parseInt(config.getValue());
+            } else if (DataTypeConst.Long.equals(config.getType())) {
+                return Long.parseLong(config.getValue());
             } else if (DataTypeConst.BOOLEAN.equals(config.getType())) {
                 return CommonConst.TRUE.equals(config.getValue());
             } else {
